+++ docx-wasm/src/delete.rs
@@ -19,12 +19,12 @@ impl Delete {
 #[wasm_bindgen]
 impl Delete {
     pub fn author(mut self, author: String) -> Delete {
-        self.0.author = author;
+        self.0 = self.0.author(author);
         self
     }
 
     pub fn date(mut self, date: String) -> Delete {
-        self.0.date = date;
+        self.0 = self.0.date(date);
         self
     }
 }