+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## docx-wasm@0.0.278-rc23 (21. Aug, 2023)
+
+- escape author in del/ins.
+
 ## docx-wasm@0.0.278-rc19 (9. Aug, 2023)
 
 - use i32 for line instead of u32.